refactor: modificando service pedido, produto e usuario com inclusao de try catch

+++ backend/search-organic/src/modelo/Produto.java
@@ -117,7 +117,7 @@ public class Produto implements Impressao {
         return taxa;
     }
 
-    public void setTaxa(float taxa) {
+    public void setTaxa(double taxa) {
         this.taxa = taxa;
     }
